Align graphite modules with the C++ headers
