+++ skia-safe/examples/graphite_readback.rs
@@ -21,10 +21,8 @@ fn main() {
     use objc2::rc::Retained;
     use objc2_metal::{MTLCreateSystemDefaultDevice, MTLDevice};
 
-    use skia_safe::{
-        AlphaType, Color4f, ColorType, ImageInfo, Paint, Rect,
-        graphite::{self, mtl as gmtl},
-    };
+    use skia_safe::gpu::graphite::{self, mtl as gmtl};
+    use skia_safe::{AlphaType, Color4f, ColorType, ImageInfo, Paint, Rect, gpu};
 
     const W: i32 = 64;
     const H: i32 = 64;
@@ -37,7 +35,8 @@ fn main() {
     let queue_ptr = Retained::as_ptr(&queue) as *mut c_void;
 
     let backend = unsafe { gmtl::BackendContext::new(device_ptr, queue_ptr) };
-    let mut context = gmtl::make_context(&backend, None).expect("make_context returned None");
+    let mut context =
+        gmtl::context_factory::make_metal(&backend, None).expect("make_metal returned None");
     let mut recorder = context
         .make_recorder(None)
         .expect("make_recorder returned None");
@@ -46,7 +45,7 @@ fn main() {
     let mut surface = graphite::surfaces::render_target(
         &mut recorder,
         &image_info,
-        graphite::Mipmapped::No,
+        gpu::Mipmapped::No,
         None,
         Some("graphite_readback"),
     )
