+++ skia-safe/src/gpu.rs
@@ -1,14 +1,20 @@
 #[cfg(feature = "d3d")]
 pub mod d3d;
+#[cfg(feature = "gpu")]
 pub mod ganesh;
 #[cfg(feature = "gl")]
 pub mod gl;
+#[cfg(feature = "graphite")]
+pub mod graphite;
+#[cfg(feature = "gpu")]
 mod mutable_texture_state;
+#[cfg(any(feature = "gpu", feature = "graphite"))]
 mod types;
 #[cfg(feature = "vulkan")]
 pub mod vk;
 
 // Ganesh re-exports (these will probably be conflict with future graphite types)
+#[cfg(feature = "gpu")]
 pub use ganesh::{
     BackendAPI, BackendFormat, BackendRenderTarget, BackendSemaphore, BackendTexture,
     DirectContext, DirectContextId, DriverBugWorkarounds, FlushInfo, PurgeResourceOptions,
@@ -16,7 +22,9 @@ pub use ganesh::{
     YUVABackendTextureInfo, YUVABackendTextures, context_options::ContextOptions, images,
 };
 
+#[cfg(feature = "gpu")]
 pub use mutable_texture_state::*;
+#[cfg(any(feature = "gpu", feature = "graphite"))]
 pub use types::*;
 
 #[cfg(feature = "metal")]
@@ -24,12 +32,14 @@ pub mod mtl {
     pub use super::ganesh::mtl::{BackendContext, types::*};
 }
 
+#[cfg(feature = "gpu")]
 pub mod surfaces {
     #[cfg(feature = "metal")]
     pub use super::ganesh::mtl::surface_metal::*;
     pub use super::ganesh::surface_ganesh::*;
 }
 
+#[cfg(feature = "gpu")]
 pub mod backend_formats {
     #[cfg(feature = "d3d")]
     pub use super::ganesh::d3d::backend_formats::*;
@@ -41,6 +51,7 @@ pub mod backend_formats {
     pub use super::ganesh::vk::backend_formats::*;
 }
 
+#[cfg(feature = "gpu")]
 pub mod backend_textures {
     #[cfg(feature = "d3d")]
     pub use super::ganesh::d3d::backend_textures::*;
@@ -52,6 +63,7 @@ pub mod backend_textures {
     pub use super::ganesh::vk::backend_textures::*;
 }
 
+#[cfg(feature = "gpu")]
 pub mod backend_render_targets {
     #[cfg(feature = "d3d")]
     pub use super::ganesh::d3d::backend_render_targets::*;
@@ -63,6 +75,7 @@ pub mod backend_render_targets {
     pub use super::ganesh::vk::backend_render_targets::*;
 }
 
+#[cfg(feature = "gpu")]
 pub mod backend_semaphores {
     #[cfg(feature = "d3d")]
     pub use super::ganesh::d3d::backend_semaphores::*;
@@ -70,6 +83,7 @@ pub mod backend_semaphores {
     pub use super::ganesh::vk::backend_semaphores::*;
 }
 
+#[cfg(feature = "gpu")]
 pub mod direct_contexts {
     #[cfg(feature = "d3d")]
     pub use super::ganesh::d3d::direct_contexts::*;
@@ -95,7 +109,7 @@ pub mod interfaces {
     pub use super::ganesh::gl::make_win_interface::interfaces::*;
 }
 
-#[cfg(test)]
+#[cfg(all(test, feature = "gpu"))]
 mod tests {
     use super::{DirectContext, RecordingContext};
 
@@ -118,3 +132,53 @@ mod tests {
         }
     }
 }
+
+#[allow(unknown_lints, clippy::uninhabited_references)]
+#[cfg(not(feature = "gpu"))]
+mod stubs {
+    use std::{
+        ops::{Deref, DerefMut},
+        ptr,
+    };
+
+    use crate::prelude::*;
+
+    #[derive(Debug)]
+    pub enum RecordingContext {}
+
+    impl NativePointerOrNullMut for Option<&mut RecordingContext> {
+        type Native = skia_bindings::GrRecordingContext;
+
+        fn native_ptr_or_null_mut(&mut self) -> *mut skia_bindings::GrRecordingContext {
+            ptr::null_mut()
+        }
+    }
+
+    #[derive(Debug)]
+    pub enum DirectContext {}
+
+    impl Deref for DirectContext {
+        type Target = RecordingContext;
+
+        fn deref(&self) -> &Self::Target {
+            unsafe { transmute_ref(self) }
+        }
+    }
+
+    impl DerefMut for DirectContext {
+        fn deref_mut(&mut self) -> &mut Self::Target {
+            unsafe { transmute_ref_mut(self) }
+        }
+    }
+
+    impl NativePointerOrNullMut for Option<&mut DirectContext> {
+        type Native = skia_bindings::GrDirectContext;
+
+        fn native_ptr_or_null_mut(&mut self) -> *mut skia_bindings::GrDirectContext {
+            ptr::null_mut()
+        }
+    }
+}
+
+#[cfg(not(feature = "gpu"))]
+pub use stubs::*;
