+++ skia-safe/src/gpu/graphite/graphite_types.rs
@@ -1,11 +1,6 @@
 use skia_bindings as sb;
 
 use crate::prelude::*;
-// Re-export the shared `skgpu` types through `gpu` rather than aliasing the
-// raw bindings again: `BackendApi` / `Mipmapped` are the same native types, and
-// `Budgeted` gets `gpu`'s bool-newtype wrapper instead of a second, name-clashing
-// incompatible type.
-pub use crate::gpu::{BackendApi, Budgeted, Mipmapped};
 
 /// Status of recording insertion (`skgpu::graphite::InsertStatus::V`).
 ///
@@ -16,46 +11,15 @@ pub use crate::gpu::{BackendApi, Budgeted, Mipmapped};
 pub use sb::skgpu_graphite_InsertStatus_V as InsertStatus;
 variant_name!(InsertStatus::Success);
 
-/// Configuration for recorder creation
-pub type RecorderOptions = Handle<sb::skgpu_graphite_RecorderOptions>;
-
-impl NativeDrop for sb::skgpu_graphite_RecorderOptions {
-    fn drop(&mut self) {
-        unsafe { sb::C_RecorderOptions_destruct(self) }
-    }
-}
-
-impl std::fmt::Debug for RecorderOptions {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("RecorderOptions").finish()
-    }
-}
-
-impl Default for RecorderOptions {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl RecorderOptions {
-    /// Create new recorder options with the C++ defaults (e.g. a 256 MiB GPU
-    /// budget). Placement-constructed rather than zero-initialized, because
-    /// `RecorderOptions` has a non-trivial constructor and members (an `sk_sp`,
-    /// a `std::optional`, and a non-zero default budget).
-    pub fn new() -> Self {
-        Self::construct(|options| unsafe { sb::C_RecorderOptions_Construct(options) })
-    }
-}
-
 /// Information for inserting a recording into the context.
 ///
-/// Borrows the [`Recording`](crate::graphite::Recording) it references (it stores
+/// Borrows the [`Recording`](super::Recording) it references (it stores
 /// a raw `fRecording` pointer), so the borrow checker keeps the `Recording` alive
 /// for as long as this info — and any `insert_recording` call using it — is in use.
 #[derive(Debug)]
 pub struct InsertRecordingInfo<'a> {
     inner: std::ptr::NonNull<sb::skgpu_graphite_InsertRecordingInfo>,
-    _recording: std::marker::PhantomData<&'a mut crate::graphite::Recording>,
+    _recording: std::marker::PhantomData<&'a mut super::Recording>,
 }
 
 impl Drop for InsertRecordingInfo<'_> {
@@ -70,7 +34,7 @@ impl<'a> InsertRecordingInfo<'a> {
     /// Takes the recording as `&mut` because `Context::insert_recording`
     /// mutates it through the stored pointer (it sets failure results for
     /// finished procs and deinstantiates volatile lazy proxies).
-    pub fn new(recording: &'a mut crate::graphite::Recording) -> Self {
+    pub fn new(recording: &'a mut super::Recording) -> Self {
         // `InsertRecordingInfo` is not POD — `fSimulatedStatus` is an
         // `InsertStatus` holding a `std::string` — and a libstdc++
         // `std::string` in SSO state points into itself, so the struct cannot
@@ -135,14 +99,6 @@ variant_name!(SyncToCpu::Yes);
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_recorder_options_creation() {
-        let options = RecorderOptions::new();
-        let _default_options = RecorderOptions::default();
-        // Should not panic and should create valid options
-        let _ = format!("{:?}", options);
-    }
 
     #[test]
     fn test_submit_info_creation() {
